Update MigrationGenerator.php
This change enables you to enter a decimal[10-2], using a hyphen instead of a comma, this will then result in "$table->decimal('amount', 10, 2);" being generated.I added an internal loop to the addColumn($field) function with an additional preg_match to catch the hyphen separated parameters. The hyphen is then reverted back to a comma within the same function.